refactor: move `started` and `key` outside storage
