+++ src/lib/storage.ts
@@ -1,3 +1,5 @@
+export const key = 'stopwatch';
+
 export interface Stopwatch{
     /// when stopwatch is paused
     timestamp: number;
@@ -5,12 +7,8 @@ export interface Stopwatch{
     duration: number;
 }
 
-export interface State{
-    /// key for localStorage
-    key: string;
+export interface Storage{
     /// number of seconds to increment on each IRL second
     increment: number;
-    /// set of names
-    started: Set<string>;
     stopwatches: { [name: string]: Stopwatch };
 }
